Booking-Backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pymongo==4.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic[email]==2.9.2
fastapi
uvicorn[standard]
pydantic
pydantic-settings
motor
pymongo
PyJWT
passlib[bcrypt]
bcrypt
python-dotenv
motor>=3.0.0